отрефакторил код - вынес безнес логику на сервере в папку comtrollers, поменял роуты api
